Move debug GUI functions from ImGui integration module to new separate Debug module
